refactor(deps): extract identity/semver helpers to identity.py
main tightened the file-length guardrail from 2450 to 2100 lines (Stage 1);
after merging main, reference.py was 2191 lines. Move the pure, stateless
identity helpers (build_dependency_unique_key, build_canonical_dependency_string,
_path_segment_pattern, the semver-range guards + segment-pattern constants) into
a sibling models/dependency/identity.py and re-export them from reference.py so
deps/lockfile.py and install/phases/resolve.py import sites stay unchanged.
Behavior-preserving; reference.py now 2099 lines (under guardrail).
